+++ src/libaevol/mutation/Inversion.cpp
@@ -68,10 +68,10 @@ void Inversion::load(gzFile backup_file) {
 }
 
 void Inversion::generic_description_string(char* str) const {
-  sprintf(str, "%" PRId8 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
+  sprintf(str, "%" PRId32 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
       " %" PRId8 " %" PRId16 " %" PRId16 " %" PRId32 " %" PRId32,
           mut_type(), pos1(), pos2(), -1, -1,
-          -1, align_score_, -1, length_, -1);
+          (int8_t) -1, align_score_, (int16_t) -1, length_, -1);
 }
 
 // ============================================================================
